Add Custom Dialog Process

+++ app/src/main/java/com/idn/ninanadia/weatherapp/models/Sys.kt
@@ -13,9 +13,9 @@ data class Sys (
     val country: String,
 
     @field:SerializedName("sunrise")
-    val sunrise: Int,
+    val sunrise: Long,
 
     @field:SerializedName("sunset")
-    val sunset: Int
+    val sunset: Long
 
     )
